+++ logtape/filesink.node.ts
@@ -1,4 +1,6 @@
-import fs from "node:fs";
+// @ts-ignore: a trick to avoid module resolution error on non-Node.js environ
+import fsMod from "./fs.ts";
+import type fsType from "node:fs";
 import { webDriver } from "./filesink.web.ts";
 import {
   type FileSinkOptions,
@@ -9,19 +11,24 @@ import {
   type Sink,
 } from "./sink.ts";
 
+// @ts-ignore: a trick to avoid module resolution error on non-Node.js environ
+const fs = fsMod as (typeof fsType | null);
+
 /**
  * A Node.js-specific file sink driver.
  */
-export const nodeDriver: RotatingFileSinkDriver<number> = {
-  openSync(path: string) {
-    return fs.openSync(path, "a");
-  },
-  writeSync: fs.writeSync,
-  flushSync: fs.fsyncSync,
-  closeSync: fs.closeSync,
-  statSync: fs.statSync,
-  renameSync: fs.renameSync,
-};
+export const nodeDriver: RotatingFileSinkDriver<number | void> = fs == null
+  ? webDriver
+  : {
+    openSync(path: string) {
+      return fs.openSync(path, "a");
+    },
+    writeSync: fs.writeSync,
+    flushSync: fs.fsyncSync,
+    closeSync: fs.closeSync,
+    statSync: fs.statSync,
+    renameSync: fs.renameSync,
+  };
 
 /**
  * Get a file sink.